Restructured some code
