+++ Social Media/Extensions.swift
@@ -117,3 +117,27 @@ extension UITableView {
         return tableView
     }
 }
+
+class UITextFieldWithPadding: UITextField {
+    
+    // MARK: - Subviews
+    
+    var textPadding = UIEdgeInsets(
+        top: 0,
+        left: 10,
+        bottom: 0,
+        right: 0
+    )
+    
+    // MARK: - Lifecycle
+    
+    override func textRect(forBounds bounds: CGRect) -> CGRect {
+        let rect = super.textRect(forBounds: bounds)
+        return rect.inset(by: textPadding)
+    }
+
+    override func editingRect(forBounds bounds: CGRect) -> CGRect {
+        let rect = super.editingRect(forBounds: bounds)
+        return rect.inset(by: textPadding)
+    }
+}
