--- Social Media/Modules/Profile/UITextFieldWithPadding.swift
@@ -1,32 +1,0 @@
-//
-//  TextFieldWithPadding.swift
-//  iOS_UI_HW3
-//
-//  Created by Philipp Lazarev on 31.05.2023.
-//
-
-import UIKit
-
-class UITextFieldWithPadding: UITextField {
-    
-    // MARK: - Subviews
-    
-    var textPadding = UIEdgeInsets(
-        top: 0,
-        left: 10,
-        bottom: 0,
-        right: 0
-    )
-    
-    // MARK: - Lifecycle
-    
-    override func textRect(forBounds bounds: CGRect) -> CGRect {
-        let rect = super.textRect(forBounds: bounds)
-        return rect.inset(by: textPadding)
-    }
-
-    override func editingRect(forBounds bounds: CGRect) -> CGRect {
-        let rect = super.editingRect(forBounds: bounds)
-        return rect.inset(by: textPadding)
-    }
-}
